feat(cmake): centralize compiler warnings with ROS2MedkitWarnings module
Add ROS2MedkitWarnings.cmake with production-grade warning flags
aligned with OpenSSF, AUTOSAR C++14, and MISRA C++ guidelines.
Enable -Werror by default. Replace per-package add_compile_options
blocks with a single include(ROS2MedkitWarnings) across all 12
packages.

New flags: -Wswitch-enum, -Wfloat-equal, -Wcast-qual, -Wundef,
-Wzero-as-null-pointer-constant, -Wextra-semi, -Wdouble-promotion,
-Wsign-conversion, -Wnon-virtual-dtor, -Wold-style-cast,
-Woverloaded-virtual, -Wcast-align, -Wnull-dereference,
-Wimplicit-fallthrough, -Wformat=2, -Wuseless-cast (GCC),
-Wduplicated-cond, -Wduplicated-branches, -Wlogical-op (GCC).

Fix all warnings surfaced by the new flags: implicit conversions,
useless casts, incomplete switch-enum coverage, float-to-double in
printf, char sign conversions.

Exclude vendored dynmsg and gtest/gmock from strict warnings via
set_source_files_properties and ros2_medkit_relax_vendor_warnings().

Closes #356

+++ src/ros2_medkit_fault_manager/src/sqlite_fault_storage.cpp
@@ -363,7 +363,7 @@ bool SqliteFaultStorage::report_fault_event(const std::string & fault_code, uint
     int64_t existing_count = check_stmt.column_int64(1);
     std::string sources_json = check_stmt.column_text(2);
     std::string current_status = check_stmt.column_text(3);
-    int32_t debounce_counter = static_cast<int32_t>(check_stmt.column_int(4));
+    int32_t debounce_counter = check_stmt.column_int(4);
 
     // CLEARED faults can be reactivated by FAILED events
     bool is_reactivation = false;
